requirements.txt
python-telegram-bot==13.11
requests==2.27.1
python-dotenv==0.20.0
